requirements.txt
django==4.1.1
argon2-cffi
django-heroku
django-storages
gunicorn
whitenoise
